// webui/src/lib/gen/models/NIMBUSClassificationRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { NIMBUSClassificationRequestCurrentObjectives } from './NIMBUSClassificationRequestCurrentObjectives';
import type { NIMBUSClassificationRequestScalarizationOptions } from './NIMBUSClassificationRequestScalarizationOptions';
import type { NIMBUSClassificationRequestSolverOptions } from './NIMBUSClassificationRequestSolverOptions';
import type { ReferencePoint } from './ReferencePoint';

/**
 * Model of the request to the nimbus method.
 */
export interface NIMBUSClassificationRequest {
	problem_id: number;
	session_id?: number | null;
	parent_state_id?: number | null;
	scalarization_options?: NIMBUSClassificationRequestScalarizationOptions;
	solver?: string | null;
	solver_options?: NIMBUSClassificationRequestSolverOptions;
	preference?: ReferencePoint;
	/** The objectives used for iteration. */
	current_objectives: NIMBUSClassificationRequestCurrentObjectives;
	num_desired?: number | null;
}