// webui/src/lib/gen/models/nIMBUSSaveRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { NIMBUSSaveRequestSessionId } from './nIMBUSSaveRequestSessionId';
import type { NIMBUSSaveRequestParentStateId } from './nIMBUSSaveRequestParentStateId';
import type { SolutionInfo } from './solutionInfo';

/**
 * Request model for saving solutions from any method's state.
 */
export interface NIMBUSSaveRequest {
	problem_id: number;
	session_id?: NIMBUSSaveRequestSessionId;
	parent_state_id?: NIMBUSSaveRequestParentStateId;
	solution_info: SolutionInfo[];
}