// webui/src/lib/gen/models/IntermediateSolutionRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { IntermediateSolutionRequestScalarizationOptions } from './IntermediateSolutionRequestScalarizationOptions';
import type { IntermediateSolutionRequestSolverOptions } from './IntermediateSolutionRequestSolverOptions';
import type { SolutionInfo } from './SolutionInfo';

/**
 * Model of the request to solve intermediate solutions between two solutions.
 */
export interface IntermediateSolutionRequest {
	problem_id: number;
	session_id?: number | null;
	parent_state_id?: number | null;
	context?: string | null;
	scalarization_options?: IntermediateSolutionRequestScalarizationOptions;
	solver?: string | null;
	solver_options?: IntermediateSolutionRequestSolverOptions;
	num_desired?: number | null;
	reference_solution_1: SolutionInfo;
	reference_solution_2: SolutionInfo;
}