// webui/src/lib/gen/models/SolutionInfo.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { SolutionInfoName } from './SolutionInfoName';

/**
 * Used when we wish to reference a solution in some `StateDB` stored in the database.
 */
export interface SolutionInfo {
	/** State of the desired solution. */
	state_id: number;
	/** Index of the desired solution. */
	solution_index: number;
	/** Name to be given to the solution. Optional. */
	name?: SolutionInfoName;
}