// webui/src/lib/gen/models/solutionInfo.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DESDEO (fast)API
 * A rest API for the DESDEO framework.
 * OpenAPI spec version: 0.1.0
 */
import type { SolutionInfoName } from './solutionInfoName';

/**
 * Used when we wish to reference a solution in some `StateDB` stored in the database.
 */
export interface SolutionInfo {
	state_id: number;
	solution_index: number;
	/** Name to be given to the solution. Optional. */
	name?: SolutionInfoName;
}